Fix the "Fork me on GitGub" banner in miew-app. And reorganize code a little bit
