+++ packages/miew-app/src/components/menu/panelslist/panels/GalleryPanel/GalleryPanel.scss
