+++ packages/miew-app/src/components/menu/panelslist/panels/InfoPanel/InfoPanel.jsx
