+++ packages/miew-app/src/components/menu/panelslist/panels/LoadPanel/LoadPanel.scss
@@ -1,4 +1,4 @@
-@import "../../../App.scss";
+@import "../../../../App.scss";
 
 .load-panel {
   padding: 15px;
