+++ packages/miew-app/src/components/menu/panelslist/panels/ToolsPanel/ToolsPanel.jsx
@@ -5,7 +5,7 @@ import { BiLogIn, BiLogOut, BiRefresh } from 'react-icons/bi';
 import { RiArrowUpDownFill, RiSave3Fill, RiCameraFill } from 'react-icons/ri';
 import { BsLink45Deg, BsFileText } from 'react-icons/bs';
 
-import { showNav } from '../../../../actions';
+import { showNav } from '../../../../../actions';
 
 import './ToolsPanel.scss';
