+++ packages/miew-app/src/components/menu/panelslist/panels/ToolsPanel/ToolsPanel.scss
@@ -1,4 +1,4 @@
-@import '../../../App.scss';
+@import '../../../../App.scss';
 
 .tools {
   color: #555;
