+++ packages/miew-app/src/containers/GalleryPanelContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-import GalleryPanel from '../components/menu/panelslist/panels/GalleryPanel.jsx';
+import GalleryPanel from '../components/menu/panelslist/panels/GalleryPanel/GalleryPanel.jsx';
 
 const mapStateToProps = (state) => ({
   viewer: state.info.viewer,
