+++ packages/miew-app/src/containers/InfoPanelContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-import InfoPanel from '../components/menu/panelslist/panels/InfoPanel.jsx';
+import InfoPanel from '../components/menu/panelslist/panels/InfoPanel/InfoPanel.jsx';
 
 const mapStateToProps = (state) => ({
   complex: state.info.complex,
