+++ packages/miew-app/src/containers/RenderSettingsPanelContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-import RenderSettingsPanel from '../components/menu/panelslist/panels/RenderSettingsPanel.jsx';
+import RenderSettingsPanel from '../components/menu/panelslist/panels/RenderSettingsPanel/RenderSettingsPanel.jsx';
 
 const mapStateToProps = (state) => ({
   viewer: state.info.viewer,
