+++ packages/miew-app/src/containers/ToolsPanelContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-import ToolsPanel from '../components/menu/panelslist/panels/ToolsPanel.jsx';
+import ToolsPanel from '../components/menu/panelslist/panels/ToolsPanel/ToolsPanel.jsx';
 
 const mapStateToProps = (state) => ({
   viewer: state.info.viewer,
